Fixed Windows compiling 03/27/2021 | 22w12b3 | 0.7.116

+++ TRAP/src/Network/Sockets/TCPSocket.cpp
@@ -245,7 +245,7 @@ TRAP::Network::Socket::Status TRAP::Network::TCPSocket::Send(const void* data, c
 	}
 
 	//Loop until every byte has been sent
-	ssize_t result = 0;
+	int64_t result = 0;
 	for(sent = 0; sent < size; sent += result)
 	{
 		//Send a chunk of data
@@ -283,7 +283,7 @@ TRAP::Network::Socket::Status TRAP::Network::TCPSocket::Receive(void* data, cons
 	}
 
 	//Receive a chunk of bytes
-	const ssize_t sizeReceived = recv(GetHandle(), static_cast<char*>(data), static_cast<int>(size), flags);
+	const int64_t sizeReceived = recv(GetHandle(), static_cast<char*>(data), static_cast<int>(size), flags);
 
 	//Check the number of bytes received
 	if (sizeReceived > 0)
